// packages/nft-connector-iota-rebased/src/models/IIotaRebasedNftMetadata.ts
// Copyright 2024 IOTA Stiftung.
// SPDX-License-Identifier: Apache-2.0.

/**
 * Metadata for an Iota Rebased NFT.
 */
export interface IIotaRebasedNftMetadata {
	/**
	 * The name of the NFT.
	 */
	name: string;

	/**
	 * The description of the NFT.
	 */
	description: string;

	/**
	 * The URI of the NFT.
	 */
	uri: string;
}